+++ spec/factory_sloth/color_spec.rb
@@ -1,0 +1,6 @@
+describe FactorySloth::Color do
+  it 'works' do
+    expect(FactorySloth::Color).to receive(:tty?).and_return true
+    expect(FactorySloth::Color.yellow('foo')).to eq "\e[33mfoo\e[0m"
+  end
+end