+++ spec/fixtures/inlined_creates.rb
@@ -1,0 +1,5 @@
+describe FactorySloth do
+  it 'skips lines with multiple create calls' do
+    arr = [create(:pointless), create(:pointless), create(:pointless)]
+  end
+end